+++ src/multilayer_perceptron/mod.rs
@@ -1,1 +1,2 @@
 pub(crate) mod multilayer_perceptron;
+pub(crate) mod utils;
